+++ page-objects/pages/generalPagesBeforeLogin/LoginPage.js
@@ -7,30 +7,31 @@ const loginFailureMessages = ['Login fehlgeschlagen.', 'Login failed.'];
 const usernameInput = 'section#loginarea input[data-testid="username"]';
 const passwordInput = 'section#loginarea input[data-testid="password"]';
 const loginBtn = 'input[data-testid="submit-login"]';
+const forgotPasswordBtn = 'a.submit-pwrecovery';
+const fillEmailInputResetPassword = '.modal-body input.form-control';
+const resetPasswordBtn = '.modal-footer > button.btn-submit';
 
 const nextSectionBtn = '#nextSection';
 const wrongLoginNotificationContainer = '.notification-content';
 
-const defaultPassword = 'Schulcloud1!';
+const defaultPassword = 'Schulcloud1qa!';
 const defaultNewPassword = 'NewPwSchulcloud1!';
 const users = {
 	teachers: {
-		klaraFallUsername: 'klara.fall@schul-cloud.org',
-		klaraFallPassword: defaultPassword,
-		cordCarlUsername: 'lehrer@schul-cloud.org',
-		cordCarlPassword: defaultPassword,
+		karlHerzogUsername: 'karl.teacher.qa@schul-cloud.org',
+		karlHerzogPassword: defaultPassword,
 	},
 
 	admins: {
-		thorstenTestUsername: 'admin@schul-cloud.org',
-		thorstenTestPassword: defaultPassword,
+		kaiPreetzUsername: 'kai.admin.qa@schul-cloud.org',
+		kaiPreetzPassword: defaultPassword,
 	},
 
 	students: {
 		fritzSchmidtUsername: 'demo-schueler@schul-cloud.org',
 		fritzSchmidtPassword: 'schulcloud',
-		paulaMayerUsername: 'paula.meyer@schul-cloud.org',
-		paulaMayerPassword: defaultPassword,
+		borisWasserUsername: 'boris.wasser.qa@schul-cloud.org',
+		borisWasserPassword: defaultPassword,
 	},
 };
 
@@ -60,6 +61,15 @@ async function clickNextSectionBtn() {
 
 async function clickLoginBtn() {
 	await elementHelpers.clickAndWait(loginBtn);
+}
+
+async function clickForgotPasswordBtn() {
+	await elementHelpers.clickAndWait(forgotPasswordBtn);
+}
+
+async function FillEmailInputAndReset(email) {
+	await waitHelpers.waitAndSetValue(fillEmailInputResetPassword, email);
+	await elementHelpers.clickAndWait(resetPasswordBtn)
 }
 
 async function clickStartUsageOfSchulcloudBtn() {
@@ -108,6 +118,8 @@ module.exports = {
 	defaultPassword,
 	defaultNewPassword,
 	users,
+	clickForgotPasswordBtn,
+	FillEmailInputAndReset,
 	performLogin,
 	performLoginActions,
 	clickOnDataProtectionBoxes,